Don't dispatch into the IoContext on enable()
When dispatching into the IoContext, new is called. As Link::enable() is meant
to be lockfree we should not do that, even though the location is very small
and not likely to cause issues. Instead of dispatching directly into the
IoContext we now enable/disable the discovery service from the
RtClientStateSetter in Controller.hpp.